Struct fcm::MessageBuilder

pub struct MessageBuilder<'a> { /* fields omitted */ }

A builder to get a Message instance.

Examples

use fcm::MessageBuilder;

let mut builder = MessageBuilder::new("<FCM API Key>", "<registration id>");
builder.dry_run(true);
let message = builder.finalize();

Methods

impl<'a> MessageBuilder<'a>

pub fn new(api_key: &'a str, to: &'a str) -> Self

Get a new instance of Message. You need to supply to.

pub fn new_multi<S>(api_key: &'a str, ids: &'a [S]) -> Self where
S: Into<Cow<'a, str>> + AsRef<str>,

Get a new instance of Message. You need to supply registration ids.

pub fn registration_ids<S>(&mut self, ids: &'a [S]) -> &mut Self where
S: Into<Cow<'a, str>> + AsRef<str>,

String value to replace format specifiers in the body string.

pub fn collapse_key(&mut self, collapse_key: &'a str) -> &mut Self

Set this parameter to identify groups of messages that can be collapsed.

pub fn priority(&mut self, priority: Priority) -> &mut Self

Set the priority of the message. You can set Normal or High priorities.

Examples:

use fcm::{MessageBuilder, Priority};

let mut builder = MessageBuilder::new("<FCM API Key>", "<registration id>");
builder.priority(Priority::High);
let message = builder.finalize();

pub fn content_available(&mut self, content_available: bool) -> &mut Self

To set the content-available field on iOS

pub fn delay_while_idle(&mut self, delay_while_idle: bool) -> &mut Self

When set to true, sends the message only when the device is active.

pub fn time_to_live(&mut self, time_to_live: i32) -> &mut Self

How long (in seconds) to keep the message on FCM servers in case the device is offline. The maximum and default is 4 weeks.

pub fn restricted_package_name(
&mut self,
restricted_package_name: &'a str
) -> &mut Self

Package name of the application where the registration tokens must match.

pub fn dry_run(&mut self, dry_run: bool) -> &mut Self

When set to true, allows you to test FCM without actually sending the message.

pub fn data(&mut self, data: &dyn Serialize) -> Result<&mut Self, Error>

Use this to add custom key-value pairs to the message. This data must be handled appropriately on the client end. The data can be anything that Serde can serialize to JSON.

Examples:

use fcm::MessageBuilder;
use std::collections::HashMap;

let mut map = HashMap::new();
map.insert("message", "Howdy!");

let mut builder = MessageBuilder::new("<FCM API Key>", "<registration id>");
builder.data(&map);
let message = builder.finalize();

pub fn notification(&mut self, notification: Notification<'a>) -> &mut Self

Use this to set a Notification for the message.

Examples:

use fcm::{MessageBuilder, NotificationBuilder};

let mut builder = NotificationBuilder::new();
builder.title("Hey!");
builder.body("Do you want to catch up later?");
let notification = builder.finalize();

let mut builder = MessageBuilder::new("<FCM API Key>", "<registration id>");
builder.notification(notification);
let message = builder.finalize();

pub fn finalize(self) -> Message<'a>

Complete the build and get a Message instance