Struct fcm::MessageBuilder

pub struct MessageBuilder { /* fields omitted */ }

A builder to get a Message instance.

Examples

use fcm::MessageBuilder;

let mut builder = MessageBuilder::new("<FCM API Key>", "<registration id>");
builder.dry_run(true);
let message = builder.finalize();

Methods

impl MessageBuilder

pub fn new<S: Into<String>>(api_key: S, to: S) -> MessageBuilder

Get a new instance of Message. You need to supply either a registration id, or a topic (/topics/...).

pub fn registration_ids<S: Into<String>>(
&mut self,
ids: Vec<S>
) -> &mut MessageBuilder

Set various registration ids to which the message ought to be sent.

pub fn collapse_key<S: Into<String>>(
&mut self,
collapse_key: S
) -> &mut MessageBuilder

Set this parameter to identify groups of messages that can be collapsed.

pub fn priority(&mut self, priority: Priority) -> &mut MessageBuilder

Set the priority of the message. You can set Normal or High priorities.

Examples:

use fcm::{MessageBuilder, Priority};
 
let mut builder = MessageBuilder::new("<FCM API Key>", "<registration id>");
builder.priority(Priority::High);
let message = builder.finalize();

pub fn content_available(
&mut self,
content_available: bool
) -> &mut MessageBuilder

To set the content-available field on iOS

pub fn delay_while_idle(
&mut self,
delay_while_idle: bool
) -> &mut MessageBuilder

When set to true, sends the message only when the device is active.

pub fn time_to_live(&mut self, time_to_live: i32) -> &mut MessageBuilder

How long (in seconds) to keep the message on FCM servers in case the device is offline. The maximum and default is 4 weeks.

pub fn restricted_package_name<S: Into<String>>(
&mut self,
restricted_package_name: S
) -> &mut MessageBuilder

Package name of the application where the registration tokens must match.

pub fn dry_run(&mut self, dry_run: bool) -> &mut MessageBuilder

When set to true, allows you to test FCM without actually sending the message.

pub fn data(&mut self, data: &Serialize) -> Result<&mut MessageBuilder, Error>

Use this to add custom key-value pairs to the message. This data must be handled appropriately on the client end. The data can be anything that Serde can serialize to JSON.

Examples:

use fcm::MessageBuilder;
use std::collections::HashMap;

let mut map = HashMap::new();
map.insert("message", "Howdy!");

let mut builder = MessageBuilder::new("<FCM API Key>", "<registration id>");
builder.data(&map);
let message = builder.finalize();

pub fn notification(
&mut self,
notification: Notification
) -> &mut MessageBuilder

Use this to set a Notification for the message.

Examples:

use fcm::{MessageBuilder, NotificationBuilder};

let mut builder = NotificationBuilder::new();
builder.title("Hey!");
builder.body("Do you want to catch up later?");
let notification = builder.finalize();
 
let mut builder = MessageBuilder::new("<FCM API Key>", "<registration id>");
builder.notification(notification);
let message = builder.finalize();

pub fn finalize(self) -> Message

Complete the build and get a Message instance